requirements.txt
streamlit==1.23.1
pandas==1.3.5
matplotlib==3.5.3
mysql-connector-python==8.0.33
pymongo==4.7.3
neo4j==5.23.0
networkx==2.6.3
